Fix translation of Gloss coords to view coords and vice versa
A bit hacky for now and I'm dumb for missing these from Vic's original
work. Gloss coordinates are (0, 0) at the centre of the screen; view
coordinates have (0, 0) at the bottom left.

 * Translate the Picture returned by drawView from view to Gloss
 * Translate incoming events from Gloss to view

Fix bug causing views to be drawn too large if their x/y-min is not 0.

Enhance the 'game' view's eventHandler to toggle colours on click for
easier testing.