Use new server transport defined in Network.Server
Serenity.Network.Server defines a map from client addresses to inboxes
and outboxes. serveClients then spawns two threads to deal with input
and output on the server's socket (also part of ServerTransport). One
thread reads packets coming into the socket and distributes them to
the appropriate inboxes based on the source address. The second reads
messages from the outboxes and sends them to the appropriate clients.

The server creates the initial ServerTransport with initTransport.
This transport is then used by acceptClient to build the client map
and pass the inboxes/outboxes to the application layer to have messages
read/written to/from.