Get weapon configuration info when determining targets
Loop over each weapon in a ship configuration and, if it's a filled
slot, find the list of enemy ships that are in range of the slot the
weapon is in.

Front mounted weapons shoot forwards, side mounted shoot left and right,
and turret weapons can fire in all three directions.