Initial work on sending client name to the server
Includes a large refactor of the networking library to allow interaction
with connected clients whilst still waiting for and accepting other
connections.

This change means that the server can fork a new thread to process new
connections whilst it deals with connected clients (i.e. receiving their
names and providing them with IDs)