Keep a group of ships in formation as they move.
If all ships in a selected group are given the same move order then they
will all end up at the same point on top of each other. So, a groupOrder
transform function takes an initial order and a list of ships to give it
to, but with the ability to modify the order for each ship.

Formation during movement orders is kept by finding the centroid of the
group of ships. Each ship is told to move to the target point but with
the same offset as its current location to the centroid.