Implement the planetary capture mechanics.
Planets are now evolvable. Every step the server checks which ships are
in capturing range of a planet. If all of the ships have the same owner
then the planet starts being captured using the following rules:

 * If the planet is partially captured by the same player then increase
   the capture percentage.
 * If the planet is partially captured by another player then decrease
   the capture percentage.
 * If the planet is uncaptured by another player then switch the players

On the other hand if there are no ships present or ships of multiple
players then the following decay rules apply:

 * If the planet is not fully captured then decrease percentage captured
 * If the planet is uncaptured then reset the capturing player